requirements.txt
# Core dependencies
openai>=1.6.0
anthropic>=0.4.0
google-api-python-client>=2.100.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Database
sqlalchemy>=2.0.0

# Utils
python-dotenv>=1.0.0
langchain>=0.1.0
